This PR refactors DagsterInstance to improve code organization and maintainability by breaking down a monolithic 3,700+ line class into focused modules. The refactoring addresses a significant technical debt issue where the entire DagsterInstance implementation was contained in a single `__init__.py` file, making the codebase difficult to navigate and maintain.
The changes include:
- Moving the core DagsterInstance class to `dagster/_core/instance/instance.py`
- Extracting type definitions and enums to `dagster/_core/instance/types.py`
- Moving utility functions and constants to `dagster/_core/instance/utils.py`
- Maintaining backward compatibility through re-exports in `__init__.py`

All existing import patterns continue to work exactly as before, ensuring zero breaking changes for external code and subclasses.

Verified backward compatibility by running the full test suite with all 33 instance tests passing. Updated import paths across the codebase including dagster-airflow and dagster-snowflake libraries. Ensured `make ruff` passes with zero errors.
