+++ python_modules/libraries/dagster-airflow/dagster_airflow/dagster_job_factory.py
@@ -10,7 +10,7 @@ from dagster import (
     _check as check,
 )
 from dagster._annotations import superseded
-from dagster._core.instance import IS_AIRFLOW_INGEST_PIPELINE_STR
+from dagster._core.instance.utils import IS_AIRFLOW_INGEST_PIPELINE_STR
 from dagster._utils.tags import normalize_tags
 
 from dagster_airflow.airflow_dag_converter import get_graph_definition_args
