+++ python_modules/libraries/dagster-airflow/dagster_airflow_tests/test_dagster_pipeline_factory/test_op_execution.py
@@ -23,7 +23,7 @@ from airflow.operators.bash_operator import BashOperator  # type: ignore (airflo
 from airflow.operators.dummy_operator import DummyOperator  # type: ignore (airflow 1 compat)
 from airflow.utils.dates import days_ago
 from dagster import DagsterEventType
-from dagster._core.instance import AIRFLOW_EXECUTION_DATE_STR
+from dagster._core.instance.utils import AIRFLOW_EXECUTION_DATE_STR
 from dagster._core.test_utils import instance_for_test
 from dagster._time import get_current_datetime
 from dagster_airflow import make_dagster_job_from_airflow_dag
