+++ python_modules/libraries/dagster-airflow/dagster_airflow_tests/test_dagster_pipeline_factory/test_tags.py
@@ -7,7 +7,7 @@ from airflow.models.dag import DAG
 from airflow.operators.bash_operator import BashOperator  # type: ignore
 from airflow.utils.dates import days_ago
 from dagster import DagsterEventType
-from dagster._core.instance import AIRFLOW_EXECUTION_DATE_STR
+from dagster._core.instance.utils import AIRFLOW_EXECUTION_DATE_STR
 from dagster._core.storage.compute_log_manager import ComputeIOType
 from dagster._core.storage.local_compute_log_manager import (
     IO_TYPE_EXTENSION,
