+++ python_modules/libraries/dagster-airflow/dagster_airflow_tests/test_persistent_db/test_persistent_airflow_db.py
@@ -13,7 +13,7 @@ from dagster import (
     build_reconstructable_job,
     execute_job,
 )
-from dagster._core.instance import AIRFLOW_EXECUTION_DATE_STR
+from dagster._core.instance.utils import AIRFLOW_EXECUTION_DATE_STR
 from dagster_airflow import (
     make_dagster_definitions_from_airflow_dags_path,
     make_persistent_airflow_db_resource,
